Profile Plot of a Stack organized in to .csv matrix
This is the macro that creates a profile plot for each image in a stack and assembles all of them into a single file. Each profile plot is a column. Number of columns is the number of images in a stack plus one column. The first column represents x-axis of images. It goes from 1 to however long an image is in pixel units.