Includes image cleanup function + automatic save function
Now this macro stops the program and waits for the user to run the cleanup code on Matlab. There is no option not to do the cleanup as ImageJ pulls cleaned files in .txt format from an intermediate folder specified in the macro. If Matlab code is never executed this folder will be empty which will result in an error. This macro also automatically saves the stack profile plot .csv file into directory specified in the macro. The name of this file is the same as the name of the folder in which images are stored.